requirements.txt
Flask>=3.0.0
requests>=2.31.0
PyYAML>=6.0.1
urllib3>=2.0.0
